make package and reorder tests and examples
