+++ roipoly/__init__.py
@@ -1,0 +1,5 @@
+from roipoly.roipoly import RoiPoly
+from roipoly.roipoly import roipoly  # compatibility for old versions
+
+__version__ = '0.1.0'
+
