+++ roipoly/roipoly.py
@@ -1,0 +1,187 @@
+"""Draw polygon regions of interest (ROIs) in matplotlib images,
+similar to Matlab's roipoly function.
+
+See the file example.py for an application.
+
+Created by Joerg Doepfert 2014 based on code posted by Daniel
+Kornhauser.
+
+"""
+
+import sys
+import logging
+import warnings
+
+import numpy as np
+import matplotlib.pyplot as plt
+from matplotlib.path import Path as MplPath
+
+logger = logging.getLogger(__name__)
+
+warnings.simplefilter('always', DeprecationWarning)
+
+
+def deprecation(message):
+    warnings.warn(message, DeprecationWarning)
+
+
+class RoiPoly:
+
+    def __init__(self, fig=None, ax=None, color='b',
+                 roicolor=None, show_fig=True):
+
+        if roicolor is not None:
+            deprecation("Use 'color' instead of 'roicolor'!")
+            color = roicolor
+
+        if fig is None:
+            fig = plt.gcf()
+
+        if ax is None:
+            ax = plt.gca()
+
+        self.start_point = []
+        self.end_point = []
+        self.previous_point = []
+        self.x = []
+        self.y = []
+        self.line = None
+        self.color = color
+        self.fig = fig
+        self.ax = ax
+
+        # Mouse event callbacks
+        self.__cid1 = self.fig.canvas.mpl_connect(
+            'motion_notify_event', self.__motion_notify_callback)
+        self.__cid2 = self.fig.canvas.mpl_connect(
+            'button_press_event', self.__button_press_callback)
+
+        if show_fig:
+            self.show_figure()
+
+    @staticmethod
+    def show_figure():
+        if sys.flags.interactive:
+            plt.show(block=False)
+        else:
+            plt.show()
+
+    def get_mask(self, current_image):
+        ny, nx = np.shape(current_image)
+        poly_verts = ([(self.x[0], self.y[0])]
+                      + list(zip(reversed(self.x), reversed(self.y))))
+        # Create vertex coordinates for each grid cell...
+        # (<0,0> is at the top left of the grid in this system)
+        x, y = np.meshgrid(np.arange(nx), np.arange(ny))
+        x, y = x.flatten(), y.flatten()
+        points = np.vstack((x, y)).T
+
+        roi_path = MplPath(poly_verts)
+        grid = roi_path.contains_points(points).reshape((ny, nx))
+        return grid
+
+    def display_roi(self, **linekwargs):
+        line = plt.Line2D(self.x + [self.x[0]], self.y + [self.y[0]],
+                          color=self.color, **linekwargs)
+        ax = plt.gca()
+        ax.add_line(line)
+        plt.draw()
+
+    def get_mean_and_std(self, current_image):
+        mask = self.get_mask(current_image)
+        mean = np.mean(np.extract(mask, current_image))
+        std = np.std(np.extract(mask, current_image))
+        return mean, std
+
+    def display_mean(self, current_image, **textkwargs):
+        mean, std = self.get_mean_and_std(current_image)
+        string = "%.3f +- %.3f" % (mean, std)
+        plt.text(self.x[0], self.y[0],
+                 string, color=self.color,
+                 bbox=dict(facecolor='w', alpha=0.6), **textkwargs)
+
+    def __motion_notify_callback(self, event):
+        if event.inaxes:
+            x, y = event.xdata, event.ydata
+            if ((event.button is None or event.button == 1) and
+                    self.line is not None):
+                # Move line around
+                x_data = [self.previous_point[0], x]
+                y_data = [self.previous_point[1], y]
+                logger.debug("draw line x: {} y: {}".format(x_data, y_data))
+                self.line.set_data(x_data, y_data)
+                self.fig.canvas.draw()
+
+    def __button_press_callback(self, event):
+        if event.inaxes:
+            x, y = event.xdata, event.ydata
+            ax = event.inaxes
+            if event.button == 1 and event.dblclick is False:
+                logger.debug("Received single left mouse button click")
+                if self.line is None:  # If there is no line, create a line
+                    self.line = plt.Line2D([x, x], [y, y],
+                                           marker='o', color=self.color)
+                    self.start_point = [x, y]
+                    self.previous_point = self.start_point
+                    self.x = [x]
+                    self.y = [y]
+
+                    ax.add_line(self.line)
+                    self.fig.canvas.draw()
+                    # Add a segment
+                else:
+                    # If there is a line, create a segment
+                    x_data = [self.previous_point[0], x]
+                    y_data = [self.previous_point[1], y]
+                    logger.debug(
+                        "draw line x: {} y: {}".format(x_data, y_data))
+                    self.line = plt.Line2D(x_data, y_data,
+                                           marker='o', color=self.color)
+                    self.previous_point = [x, y]
+                    self.x.append(x)
+                    self.y.append(y)
+
+                    event.inaxes.add_line(self.line)
+                    self.fig.canvas.draw()
+
+            elif (((event.button == 1 and event.dblclick is True) or
+                   (event.button == 3 and event.dblclick is False)) and
+                  self.line is not None):
+                # Close the loop and disconnect
+                logger.debug("Received single right mouse button click or "
+                             "double left click")
+                self.fig.canvas.mpl_disconnect(self.__cid1)
+                self.fig.canvas.mpl_disconnect(self.__cid2)
+
+                self.line.set_data([self.previous_point[0],
+                                    self.start_point[0]],
+                                   [self.previous_point[1],
+                                    self.start_point[1]])
+                ax.add_line(self.line)
+                self.fig.canvas.draw()
+                self.line = None
+
+                if sys.flags.interactive:
+                    pass
+                else:
+                    #  Figure has to be closed so that code can continue
+                    plt.close(self.fig)
+
+    # For compatibility with old version
+    def displayMean(self, *args, **kwargs):
+        deprecation("Use 'display_mean' instead of 'displayMean'!")
+        return self.display_mean(*args, **kwargs)
+
+    def getMask(self, *args, **kwargs):
+        deprecation("Use 'get_mask()' instead of 'getMask'!")
+        return self.get_mask(*args, **kwargs)
+
+    def displayROI(self, *args, **kwargs):
+        deprecation("Use 'display_roi' instead of 'displayROI'!")
+        return self.display_roi(*args, **kwargs)
+
+
+# For compatibility with old version
+def roipoly(*args, **kwargs):
+    deprecation("Import 'RoiPoly' instead of 'roipoly'!")
+    return RoiPoly(*args, **kwargs)
